requirements.txt
ollama
google-api-python-client
google-auth-oauthlib
bs4